pyproject.toml
[tool.poetry]
name = "ragms02"
version = "0.1.0"
description = "File-watcher sidecar and RAG LLM Interface Service"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{ include = "ragms02", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
fastapi = "*"
uvicorn = "*"
watchdog = "*"
sqlite-vss = "*"
requests = "*"
pydantic = "*"
ulid-py = "*"
sphinx = "*"
pytest = "*"
httpx = "*"
numpy = "^1.24"
langchain = "^0.3.26"
tiktoken = "^0.9.0"
google-generativeai = "^0.8.5"

[tool.poetry.group.dev.dependencies]
black = "*"
flake8 = "*"
pytest = "*"
pytest-cov = "*"
pre-commit = "*"
sphinx-autodoc-typehints = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
